Update pom.xml and refactor database-related classes
Bumped the version of MYSQL-Api in pom.xml to 5.7.2. Refactored code for better readability and efficiency in Database.java, MYSQL.java, and other related classes. The modified queries used in the executeQuery() method of QueryHelper interface now return a list of results, addressing previous issues when more than one row was queried.