Update version and add connection test query
The version of MYSQL-Api in pom.xml is updated from 5.7.5 to 5.7.6. Additionally, a connection test query is added in MYSQL.java to test the validity of a connection by executing a query that most probably be valid on the database. These changes ensure project dependencies are up to date and increase the robustness of the database connection handling.